+++ config/configHandler.py
@@ -2,19 +2,24 @@ import configparser
 import sys
 import os
 from os import path
+import shutil
 import logging
+
+from time import time
+from datetime import datetime
 
 import config.dbConstants as dbConst
 
 
 class DatabaseConfiguration:
     def __init__(self):
-        # Read the ini file from local dir
-        # 1.6.7 - if default config file does not exist, read default_config.ini file instead
         db_config = configparser.ConfigParser()
+
+        # Attempt to read the config.ini file from local dir
         if path.exists(dbConst.CONFIG_FILE):
             print('ConfigHandler - reading from config.ini file')
             db_config.read(dbConst.CONFIG_FILE)
+        # 1.6.7 - if default config file does not exist, read default_config.ini file instead
         else:
             print('ConfigHandler - Config.ini not found, generating for the first time')
             db_config.read(dbConst.DEFAULT_CONFIG_FILE)
@@ -29,7 +34,7 @@ class DatabaseConfiguration:
 
         # Set the logger - have to import logger settings first
         logging.basicConfig(stream=sys.stdout, level=self.logLevel[dbConst.CONF_HANDLER])
-        logger = logging.getLogger(dbConst.CONF_HANDLER)
+        self.logger = logging.getLogger(dbConst.CONF_HANDLER)
 
         # Set version
         self.codeVersion = db_config['DEFAULT']['codeVersion']
@@ -45,39 +50,32 @@ class DatabaseConfiguration:
         self.typeColumnIndex = self.get_column(dbConst.TYPE)
 
         # Debug statements
-        logger.info('Initializing ConfigHandler')
-        logger.debug('DEBUG - DB delimiter:' + db_config.get('DEFAULT', 'delimiter'))
-        logger.debug('DEBUG - DB format:' + db_config.get('DEFAULT', 'format'))
-        logger.debug('DEBUG - DB filename:' + db_config.get('DEFAULT', 'filename'))
-        logger.debug('DEBUG - Config sections:' + str(db_config.sections()))
+        self.logger.info('Initializing ConfigHandler')
+        self.logger.debug('DEBUG - DB delimiter:' + db_config.get('DEFAULT', 'delimiter'))
+        self.logger.debug('DEBUG - DB format:' + db_config.get('DEFAULT', 'format'))
+        self.logger.debug('DEBUG - DB filename:' + db_config.get('DEFAULT', 'filename'))
+        self.logger.debug('DEBUG - Config sections:' + str(db_config.sections()))
 
     def __repr__(self):
         return 'Log levels: ' + str(self.logLevel)
 
-    def update_log_level(self, logger_name, new_log_level):
-        # Set the logger - have to import logger settings first
-        logging.basicConfig(stream=sys.stdout, level=self.logLevel[dbConst.CONF_HANDLER])
-        logger = logging.getLogger(dbConst.CONF_HANDLER)
+    # Get column by index, for parsing the DB format from config.ini file
+    def get_column(self, name):
+        try:
+            column_index = self.columns.index(name)
+            return column_index
+        except ValueError:
+            return -1
 
+    def update_log_level(self, logger_name, new_log_level):
         update_result = ''
 
         # Only update if valid logger level:
         if new_log_level in dbConst.VALID_LOG_LEVELS:
             # Only update if valid logger:
             if logger_name in self.logLevel:
-                logger.debug('Found ' + logger_name + ' in ' + str(self.logLevel))
-                # Read the ini file from local dir and update it
-                db_config = configparser.ConfigParser()
-                #with open('config/config.ini', 'r') as config_file:
-                db_config.read(dbConst.CONFIG_FILE)
-                db_config[dbConst.SECT_LOG][logger_name] = new_log_level
-                logger.debug('Read in config as: ' + str(db_config))
-                logger.debug('Config sections:' + str(db_config.sections()))
-
-                with open(dbConst.CONFIG_FILE, 'w') as config_file_updater:
-                    db_config.write(config_file_updater)
-
-                # Also update the in-memory log level
+                self.logger.debug('Found ' + logger_name + ' in ' + str(self.logLevel))
+                # Update the in-memory log level
                 self.logLevel[logger_name] = new_log_level
                 update_result = 'Logger: ' + logger_name + ' updated to log level: ' + new_log_level
             else:
@@ -89,10 +87,21 @@ class DatabaseConfiguration:
 
         return update_result
 
-    # Get column by index, for parsing the DB format from config.ini file
-    def get_column(self, name):
-        try:
-            column_index = self.columns.index(name)
-            return column_index
-        except ValueError:
-            return -1
+    # Persist the config changes, to be called upon exit
+    def persist_config(self):
+        # Backup the existing conf file first into the backup dir
+        backup_file_name = 'config-' + str(datetime.fromtimestamp(time()).strftime("%Y%m%d-%H%M%S")) + '.ini.'
+        self.logger.debug('Backup filename:' + backup_file_name)
+        # TODO - diff the files and only update if different
+        shutil.copy(dbConst.CONFIG_FILE, 'config/backup/' + backup_file_name)
+
+        # Create a configParser from the existing config.ini to use as a base to update
+        temp_db_config = configparser.ConfigParser()
+        temp_db_config.read(dbConst.CONFIG_FILE)
+        # Update the config with the in-memory logger levels:
+        for logger_name in self.logLevel:
+            self.logger.debug('Logger ' + logger_name + ' . Old:' + str(temp_db_config[dbConst.SECT_LOG][logger_name]) + ', new:' + str(self.logLevel[logger_name]))
+            temp_db_config[dbConst.SECT_LOG][logger_name] = self.logLevel[logger_name]
+
+        with open(dbConst.CONFIG_FILE, 'w') as config_file_writer:
+            temp_db_config.write(config_file_writer)